requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.39
groq==0.19.0
langchain-community==0.3.19